feat: add answerTime field in dto #23

+++ src/main/java/com/kospot/kospot/domain/game/dto/request/EndGameRequest.java
@@ -14,6 +14,6 @@ public class EndGameRequest {
         private double submittedLat;
         private double submittedLng;
         private double answerDistance;
-        private Long answerTime;
+        private double answerTime;
     }
 }
